requirements.txt
asgiref==3.3.1
Django==3.1.5
django-ninja==0.10.1
pkg-resources==0.0.0
pydantic==1.7.3
python-decouple==3.4
pytz==2021.1
sqlparse==0.4.1
